+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Deprecate {ERB,Erubis}Template#default_output_variable{,=} (jeremyevans)
 * Avoid holding mutex while compiling template methods (jeremyevans)
 * Add Tilt::StaticTemplate for templates that return the same output for every render (jeremyevans)
+* Template#prepare no longer needs to be overridden if no preparation work is needed (jeremyevans)
 
 The repository switched to https://github.com/jeremyevans/tilt, so issue references above are for that
 repository, and issue references below are for the previous repository (https://github.com/rtomayko/tilt).